Removed divs of the images and replaced them with a list array iList (JS structural changes and CSS changes)
Added in the list array for the images.

Edited slideImages var and changed to slide

Changed the next , previous, slideleft and slideright, show, hide functions so that the JS now edits the class to flick through the pictures.

Also increased the gap of between switches through the list array to 3 seconds so I can test the transition (which isn't currently displaying).

Removed the slide class